+++ src/stLFR/TagId.h
@@ -19,19 +19,19 @@ namespace BGIQD {
                 TagId & operator =(const TagId & );
             public:
                 // Add a tag and return it's number.
-                long AddTag(const std::string & tag );
+                long long AddTag(const std::string & tag );
                 // Return the number of a tag , -1 for not found .
-                long  Id( const std::string & tag ) const ;
+                long long  Id( const std::string & tag ) const ;
 
                 // Assign a number for a tag .
-                void AssignTag( const std::string & tag , long number );
+                void AssignTag( const std::string & tag ,long long number );
 
                 // Print barcodeList into file
                 //  format is :
                 //      "tag\tid\n"
                 void Print(std::ostream & ost ) const;
             private:
-                std::map<std::string , long > m_tag2num;
+                std::map<std::string ,long  long > m_tag2num;
                 //std::map<long , std::string > m_num2tag;
                 long curr ;
         };
